requirements.txt
# Web Framework & Pydantic
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-multipart>=0.0.9
httpx>=0.27.0

# Database & Caching
sqlalchemy[asyncio]>=2.0.0
alembic>=1.13.0
asyncpg>=0.29.0
redis[asyncio]>=5.0.0

# Vector DB & Search
qdrant-client[fastembed]>=1.8.0
rank-bm25>=0.2.2

# AI/ML - Open Source Only
transformers>=4.38.0
torch>=2.2.0
sentence-transformers>=2.2.0
accelerate>=0.27.0
bitsandbytes>=0.42.0

# PDF Parsing & Extraction
pymupdf>=1.24.0
pytesseract>=0.3.10
pdf2image>=1.16.0
unstructured[all-docs]>=0.12.0

# RAG & LLM Framework (Alternative to llama-cloud)
llama-index>=0.10.0
llama-index-llms-huggingface>=0.1.0
llama-index-embeddings-huggingface>=0.1.0

# Utilities
loguru>=0.7.0
tqdm>=4.66.0
python-dotenv>=1.0.0

